+++ CK.MQTT.Client/PublishPackets/SmallOutgoingApplicationMessage.cs
@@ -18,8 +18,8 @@ namespace CK.MQTT
         {
             _memory = payload;
         }
-        
-        protected override int PayloadSize => _memory.Length;
+
+        protected override uint PayloadSize => (uint)_memory.Length;
 
         protected async override ValueTask<WriteResult> WritePayloadAsync( PipeWriter pw, CancellationToken cancellationToken )
         {
